+++ src/image.cpp
@@ -406,7 +406,7 @@ namespace Exiv2 {
                 if ( allocate > (long long) io.size() ) {
                     throw Error(57);
                 }
-                DataBuf  buf(allocate);  // allocate a buffer
+                DataBuf  buf((long)allocate);  // allocate a buffer
                 std::memcpy(buf.pData_,dir.pData_+8,4);  // copy dir[8:11] into buffer (short strings)
                 const bool bOffsetIsPointer = count*size > 4;
